+++ packages/tree/src/SerializableTree.ts
@@ -3,8 +3,7 @@ import { ATree } from './ATree.js';
 import { SerializableTreeObject } from './SerializableTreeObject.js';
 import { SerializableTreeNode, TreeNodeJSON } from './TreeNode/SerializableTreeNode.js';
 
-export class SerializableTree<V extends SerializableTreeObject> extends ATree<V, SerializableTreeNode<V>, 'SerializableTree'>
-  implements JSONable<TreeNodeJSON> {
+export class SerializableTree<V extends SerializableTreeObject> extends ATree<V, SerializableTreeNode<V>, 'SerializableTree'> implements JSONable<TreeNodeJSON> {
   public readonly noun: 'SerializableTree' = 'SerializableTree';
 
   public static of<VT extends SerializableTreeObject>(root: SerializableTreeNode<VT>): SerializableTree<VT> {