+++ packages/collection/CHANGELOG.md
@@ -1,26 +1,54 @@
-# [@jamashita/lluvia-collection-v2.0.0](https://github.com/jamashita/lluvia/compare/@jamashita/lluvia-collection-v1.8.0...@jamashita/lluvia-collection-v2.0.0) (2021-07-21)
+# [@jamashita/lluvia-collection-v2.2.0](https://github.com/jamashita/lluvia/compare/@jamashita/lluvia-collection-v2.1.0...@jamashita/lluvia-collection-v2.2.0) (2021-08-14)
 
 
 ### Features
 
-* 🎸 update ([6244c09](https://github.com/jamashita/lluvia/commit/6244c095b2de0af6df96da944b86dc6f3bc38285))
+* 🎸 update ([2b2cddc](https://github.com/jamashita/lluvia/commit/2b2cddc970d387683bb03457b0ed4e1b66109a74))
+* 🎸 update ([11be8f0](https://github.com/jamashita/lluvia/commit/11be8f07541ac6d25fef570a6ebac4df104cc712))
+
+# [@jamashita/lluvia-collection-v2.1.0](https://github.com/jamashita/lluvia/compare/@jamashita/lluvia-collection-v2.0.0...@jamashita/lluvia-collection-v2.1.0) (2021-07-22)
+
+
+### Features
+
+* 🎸 update ([4fa746b](https://github.com/jamashita/lluvia/commit/4fa746b62a84f19cd3f0f066c00b460752a1a55c))
+
+# Change Log
+
+All notable changes to this project will be documented in this file.
+See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
+
+# [2.1.0](https://github.com/jamashita/lluvia.git/packages/collection/compare/@jamashita/lluvia-collection@1.8.0...@jamashita/lluvia-collection@2.1.0) (2021-07-22)
+
+
+### Features
+
+* 🎸 update ([6244c09](https://github.com/jamashita/lluvia.git/packages/collection/commit/6244c095b2de0af6df96da944b86dc6f3bc38285))
 
 
 ### BREAKING CHANGES
 
 * 🧨 new ver
 
-# [@jamashita/lluvia-collection-v1.8.0](https://github.com/jamashita/lluvia/compare/@jamashita/lluvia-collection-v1.7.0...@jamashita/lluvia-collection-v1.8.0) (2021-07-21)
+
+
+
+
+## [2.0.1](https://github.com/jamashita/lluvia.git/packages/collection/compare/@jamashita/lluvia-collection@1.8.0...@jamashita/lluvia-collection@2.0.1) (2021-07-22)
 
 
 ### Features
 
-* 🎸 update ([780d8b4](https://github.com/jamashita/lluvia/commit/780d8b4b19124ac1758b23105cddb13f500b5493))
+* 🎸 update ([6244c09](https://github.com/jamashita/lluvia.git/packages/collection/commit/6244c095b2de0af6df96da944b86dc6f3bc38285))
 
-# Change Log
 
-All notable changes to this project will be documented in this file.
-See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
+### BREAKING CHANGES
+
+* 🧨 new ver
+
+
+
+
 
 # 1.8.0 (2021-07-21)